api/requirements.txt
fastapi
uvicorn
torch
transformers
faiss-cpu
biopython
tqdm
